requirements.txt
babel==2.10.3
python-dateutil==2.8.2
flask-moment== 1.0.4
flask-wtf==1.0.1
flask_sqlalchemy==2.5.1
flask-migrate==3.1.0
psycopg2-binary==2.9.3